SuperCommit
Cambios en todo lado

+++ app/models/RegionTotal.java
@@ -51,57 +51,57 @@ public class RegionTotal extends Model {
 	/**
 	 * Average gold earned in the region
 	 */
-	public int avg_gold;
+	public double avg_gold;
 	
 	/**
 	 * Average double kills in the region
 	 */
-	public int avg_double_kills;
+	public double avg_double_kills;
 	
 	/**
 	 * Average triple kills in the region
 	 */
-	public int avg_triple_kills;
+	public double avg_triple_kills;
 	
 	/**
 	 * Average quadra kills in the region
 	 */
-	public int avg_quadra_kills;
+	public double avg_quadra_kills;
 	
 	/**
 	 * Average penta kills in the region
 	 */
-	public int avg_penta_kills;
+	public double avg_penta_kills;
 	
 	/**
 	 * Average turrets destroyed in the region
 	 */
-	public int avg_turrets_destroyed;
+	public double avg_turrets_destroyed;
 	
 	/**
 	 * Time of the longest game in the region
 	 */
-	public int longest_game_time;
+	public double longest_game_time;
 	
 	/**
 	 * Average game time in the region
 	 */
-	public int average_game_time;
+	public double average_game_time;
 	
 	/**
 	 * Average krakens spent in the region
 	 */
-	public int average_krakens;
+	public double average_krakens;
 	
 	/**
 	 * Average dragons killed in the region
 	 */
-	public int average_dragons;
+	public double average_dragons;
 	
 	/**
 	 * Average barons killed in the region
 	 */
-	public int average_barons;
+	public double average_barons;
 	
 	/**
 	 * Brawlers statics
